Don't enforce block hash in checkpoint votes
This smells like room for premature deregistration about people in
quorums not being aware of all possible alt-chains and their children
since the default behaviour for alternative blocks is to NOT relay
them.

So some nodes may be aware of other blocks that others are not and be
unable to say that the hash belongs to a block and prematurely
decommission a node, so play it safe and assume it is so.

We should be fixing this by verifying the vote signature somehow.