Optimize control socket common case
The map lookup is only really needed here for supporting multiple
SNNetwork instances, which is not going to be a common case.  This adds
an optimization for the common singleton case (which probably incurs a
small penalty for the multiple-SNNetwork case, but we aren't even
currently using that in loki).