Remove dead code
None of this code is used or called or invoked anywhere, yet complicates
things by throwing in unneeded multiple inheritance, a unique_ptr that
is never touched, and compiling functions (and an extra .cpp file) that
are never called which are full of FIXMEs and TODOs and probably don't
work.  None of this has been touched in years, either.

Purge it.