Split up and DRY prune(); make Blockchain and tx_memory_pool Lockable
Extracts the tx removal lambda into a new method.

DRYs the nearly identical prune loop bodies (the only difference between
them is the iteration order).

Fix undefined behaviour in mempool pruning: on removal of a standard tx
the iterator was updated to the result of `.erase()`, but that's wrong
because the loop is going in backwards order, and so the first prune
would leave the iterator pointing at `.end()` which would be deferenced
in the next iteration.  The (new, DRY) lambda now iterates in the
desired direction.

This also makes (and uses) blockchain and tx_pool implement the standard
library Lockable concept by adding a `try_lock()` so that they can be
locked simultaneously.  It also updates them to hold a
boost::recursive_mutex directly rather than going through the pointless
and horribly designed epee `critical_section` wrapper (seriously who
thinks that making a mutex class copyable where a copy holds a new,
entirely unrelated mutex is a good idea?)