Don't check alt quorums when processing blocks on the mainchain
If you accept a block on the main chain due to an alternative quorum,
eventually that alternative quorum will get pruned as it corresponds to an
alternative block and not the current block.

This should not be accepted via the main chain, but instead the block
should be passed through the alt chain block added hook.