set_daemon mutex fix and simplifications
set_daemon got changed to set the long polling daemon, but then this
later got removed but the extra logic (now unnecessary) remained.  This
(mostly) reverts it back to the simpler pre-rpc-long-poll code.

The lock here also looked later than it should be: elsewhere it is
guarding changes to m_daemon_address and m_daemon_login but in the
change above it was moved to not guard them anymore.